Support collection level monitoring metrics
