+++ internal/proxy/channels_mgr_test.go
@@ -370,9 +370,8 @@ func Test_singleTypeChannelsMgr_removeStream(t *testing.T) {
 			},
 		},
 	}
-	err := m.removeStream(100)
-	assert.NoError(t, err)
-	_, err = m.lockGetStream(100)
+	m.removeStream(100)
+	_, err := m.lockGetStream(100)
 	assert.Error(t, err)
 }
 
@@ -384,8 +383,7 @@ func Test_singleTypeChannelsMgr_removeAllStream(t *testing.T) {
 			},
 		},
 	}
-	err := m.removeAllStream()
-	assert.NoError(t, err)
-	_, err = m.lockGetStream(100)
+	m.removeAllStream()
+	_, err := m.lockGetStream(100)
 	assert.Error(t, err)
 }
